requirements.txt
django
djangorestframework
pymysql
cryptography
mongoengine
django-mongoengine
psycopg2